Initial commit
This commit is the first commit for the tool. it include the python port of pathos tool. The tool is
written in Go (there is Ruby version too). I wanted to port it to python so that i modify and control it
because I don't know neither Ruby or Go. There are some problems.

1- Flickering when using arrows
2- Problems with Mac terminals that they don't move with the arrows
3- Colors are not consistent.

Authored-by: Mohamed Elashri <mail@elashri.com>